requirements.txt
pandas
sqlalchemy
sqlite3
matplotlib.pyplot
seaborn
sklearn.cluster
sklearn.preprocessing